Enhance database seeding with Faker in Prisma
The update introduces the use of the 'faker' library to seed the database for enhanced data simulation. It is crucial to note that we must refrain from using the SEED argument while running 'npx prisma db seed' to populate the database continually, as this would violate unique constraints for PKs uuid. Dropped data should be reconstructed by using 'npx prisma migrate reset'.