Update logic for `edu_institution_id` and `intern_hours_required`
The conditional check for `edu_institution_id` and `intern_hours_required` has been updated in `seedLearners` function. This change will set the `edu_institution_id` to `null` if a user is not enrolled in college, and `intern_hours_required` is set to `0` if a user is not enrolled in college, ensuring data consistency.