Update `generateSkillCategoryData` to accept category as parameter
The function `generateSkillCategoryData` in `seed.ts` has been updated to accept a category as a parameter. This change is implemented to allow more flexibility in determining the category name. Now, instead of generating a random department name as the category name, it will use the passed category as category name.