Refactor JobListingHasSkillCategory seeding function
The 'seedJobListingHasSkillCategory' function has been refactored to return an array of the created records. The function now randomizes the number of skill categories for each job listing, improving the diversity of data and realism in testing scenarios. This also provides more detailed information about the state after seeding, enhancing traceability and debugging experience.