Add Spinner component to the app
A new Spinner component is added to handle the display of loading messages in the application. It accepts props to optionally display a text message and to control the size of the spinner. The design of this component is completely customizable and it enhances the user experience during data fetching or processing.