Move seed.ts file and update schema
The 'seed.ts' file was shifted to a new location: 'lib/prisma'. Additionally, several changes were made to the schema. The relationship 'job_listing_has_skill_category' was updated and renamed to 'jobListingSkillCategories'. A new field 'skill_class' with a default value of 'none' was added to the 'skill' table. A new enumeration 'SkillClassification' was also introduced.